Compiling broken stuff into something that works.
It seems like a bunch of people had something that may have worked in the past but for whatever reason it now does not work.  Whether this was some random thing that happened or an intentional "code breaking" attempt to make the commercial version of the official pipboy the only thing for people to buy, the world may never now.  In any case, this is a fun way to waste time.